pai-web-agent/pyproject.toml
[project]
name = "pai-web-agent"
version = "0.1.0"
description = "LangGraph-based supervised agent with OpenAI and Tavily API integration"
authors = [
    {name = "PAI Team", email = "team@pai.com"}
]
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "langchain>=0.3.7",
    "langchain-openai>=0.2.8",
    "langchain-tavily>=0.2.0",
    "langgraph>=0.2.45",
    "python-dotenv>=1.0.1",
    "streamlit>=1.28.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "black>=23.0.0",
    "flake8>=6.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["."]

[tool.black]
line-length = 88
target-version = ['py39']

[tool.flake8]
max-line-length = 88
extend-ignore = ["E203", "W503"]
